+++ test/integration/types.ts
@@ -31,6 +31,7 @@ export interface SharedState {
   integrationParentRemId?: string;
   integrationParentTitle?: string;
   searchByTagTag?: string;
+  searchByTagTagRemId?: string;
   noteAId?: string;
   noteBId?: string;
   noteCId?: string;